Refactor fileset and collection streaming into behavior
Better export derivatives and full size files based on derivative status
