+++ app/services/oregon_digital/collection_streamer.rb
@@ -4,22 +4,19 @@ module OregonDigital
   # Service for streaming a collections sub-collections and works into a zip file
   class CollectionStreamer
     include Enumerable
-
-    def self.stream(collection, &chunks)
-      streamer = new(collection)
-      streamer.each(&chunks)
-    end
+    include OregonDigital::StreamingDownloadBehavior
 
     attr_reader :collection
 
-    def initialize(collection)
+    def initialize(collection, standard)
       @collection = collection
+      @standard = standard
     end
 
     def each(&chunks)
       writer = ZipTricks::BlockWrite.new(&chunks)
 
-      ZipTricks::Streamer.open(writer) do |zip|
+      ZipTricks::Streamer.open(writer, auto_rename_duplicate_filenames: true) do |zip|
         stream_collection(collection, '/', zip)
       end
     end
@@ -37,22 +34,28 @@ module OregonDigital
         file_writer << collection.csv_metadata
       end
 
-      stream_works(folder, zip)
-    end
-
-    def stream_works(folder, zip)
-      # Add files from works in this collection
+      # Add low quality works from collection
       collection.child_works.each do |work|
-        work.file_sets.each do |file_set|
-          file_name = "#{folder}#{file_set.title.first}"
-
-          zip.write_deflated_file(file_name) do |file_writer|
-            file_set.files.first.stream.each do |chunk|
-              file_writer << chunk
-            end
-          end
-        end
+        stream_works_low(work, zip, folder)
       end
     end
+
+    # def stream_works_low(work, folder = '', zip)
+    #   work.file_sets.each do |file_set|
+    #     file_name = "#{folder}#{file_set.label}"
+    #     deriv_name = derivative_name(file_set)
+
+
+    #     begin
+    #       file = File.open(Hyrax::DerivativePath.derivative_path_for_reference(file_set, deriv_name), 'rb')
+    #       zip.write_deflated_file(file_name) do |file_writer|
+    #         IO.copy_stream(file, file_writer)
+    #       end
+    #       file.close
+    #     rescue Errno::ENOENT
+    #       stream_files_from_fileset(file_set, zip)
+    #     end
+    #   end
+    # end
   end
 end
